Logically separate H5Type file creation from network construction
